correctly match the first and last words
If the first or last word was a small word and also used elsewhere
in the title, it would be incorrectly capitalized.

For example, "the quick and the dead" was incorrectly transformed to
"The Quick and The Dead". It is now "The Quick and the Dead".

Fixes #13. Thanks to Jason Rosendale.